delete, update, proxy

+++ app/api/item/delete/[id]/route.ts
@@ -1,13 +1,20 @@
 import supabase from "@/app/utils/database";
 import { NextResponse } from "next/server";
 
-export const DELETE = async (_, context : any) => {
+export const DELETE = async (request :any, context : any) => {
+  const req = await request.json;
+  const params = await context.params;
   try {
-    const params = await context.params;
-    const {error} = await supabase.from("items").delete().eq("id", params.id);
+    const {data, error} = await supabase.from("items").select().eq("id",params.id).single();
     if(error)throw new Error(error.message);
-    return NextResponse.json({message: "delete item"});
+    if(data.email === req.email){
+      const {error} = await supabase.from("items").delete().eq("id", params.id);
+      if(error)throw new Error(error.message);
+      return NextResponse.json({message: "delete item"});
+    }else{
+      return NextResponse.json({message: "ほかの人が作成したアイテムです"});
+    }
   } catch (error) {
-    return NextResponse.json({message: "falied to delete item"});
+    return NextResponse.json({message: `falied to delete item :${error}`});
   }
 }
